@@ -2,18 +2,15 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
 using System;
-using System.Collections.Immutable;
 using System.IO;
 using System.Linq;
 using System.Reflection;
-using System.Runtime.CompilerServices;
-using System.Threading.Tasks;
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp;
 using Microsoft.CodeAnalysis.CSharp.Testing;
-using Microsoft.CodeAnalysis.Diagnostics;
 using Microsoft.CodeAnalysis.Testing.Verifiers;
 using Microsoft.CodeAnalysis.Text;
+using Xunit;
 
 public static partial class CSharpCodeFixVerifier<TAnalyzer, TCodeFix>
 {
@@ -26,7 +23,8 @@ public static partial class CSharpCodeFixVerifier<TAnalyzer, TCodeFix>
 
             this.SolutionTransforms.Add((solution, projectId) =>
             {
-                var parseOptions = (CSharpParseOptions)solution.GetProject(projectId).ParseOptions;
+                var parseOptions = (CSharpParseOptions?)solution.GetProject(projectId)?.ParseOptions;
+                Assert.NotNull(parseOptions);
                 solution = solution.WithProjectParseOptions(projectId, parseOptions.WithLanguageVersion(LanguageVersion.CSharp7_3));
 
                 return solution;
